@@ -18,16 +18,17 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 2.0.1+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  card_swiper: ^1.0.2
+  cupertino_icons: ^1.0.3
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  flutter_rating_bar: ^3.0.1+1
-  flutter_swiper: ^1.1.6
-  google_fonts: ^1.1.0
-  shimmer: ^1.1.1
+  flutter_lints: ^1.0.4
+  flutter_rating_bar: ^4.0.0
+  google_fonts: ^2.1.0
+  shimmer: ^2.0.0
 
 dev_dependencies:
   flutter_test:
