@@ -5,9 +5,11 @@ import 'package:swiggy_ui/utils/ui_helper.dart';
 import 'package:swiggy_ui/widgets/custom_divider_view.dart';
 
 class GenieScreen extends StatelessWidget {
+  const GenieScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    final services = Genie.getGenieServices();
+    const services = Genie.genieServices;
 
     return Scaffold(
       body: SafeArea(
@@ -21,7 +23,7 @@ class GenieScreen extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               IconButton(
-                icon: Icon(Icons.arrow_back, color: Colors.white),
+                icon: const Icon(Icons.arrow_back, color: Colors.white),
                 onPressed: () {
                   Navigator.pop(context);
                 },
@@ -31,7 +33,7 @@ class GenieScreen extends StatelessWidget {
                   child: Column(
                     mainAxisSize: MainAxisSize.max,
                     children: <Widget>[
-                      UIHelper.verticalSpaceMedium(),
+                      UIHelper.verticalSpaceMedium,
                       Column(
                         crossAxisAlignment: CrossAxisAlignment.center,
                         children: <Widget>[
@@ -40,9 +42,12 @@ class GenieScreen extends StatelessWidget {
                             children: <Widget>[
                               Text(
                                 'Genie',
-                                style: Theme.of(context).textTheme.headline4.copyWith(color: Colors.white),
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .headline4!
+                                    .copyWith(color: Colors.white),
                               ),
-                              UIHelper.horizontalSpaceSmall(),
+                              UIHelper.horizontalSpaceSmall,
                               Image.asset(
                                 'assets/images/delivery-boy.png',
                                 height: 50.0,
@@ -50,17 +55,18 @@ class GenieScreen extends StatelessWidget {
                               )
                             ],
                           ),
-                          UIHelper.verticalSpaceExtraSmall(),
+                          UIHelper.verticalSpaceExtraSmall,
                           Text(
                             'Anything you need, delivered',
-                            style: Theme.of(context).textTheme.bodyText1.copyWith(
-                                  color: Colors.grey[200],
-                                  fontSize: 17.0,
-                                ),
+                            style:
+                                Theme.of(context).textTheme.bodyText1!.copyWith(
+                                      color: Colors.grey[200],
+                                      fontSize: 17.0,
+                                    ),
                           )
                         ],
                       ),
-                      UIHelper.verticalSpaceMedium(),
+                      UIHelper.verticalSpaceMedium,
                       Container(
                         padding: const EdgeInsets.all(20.0),
                         decoration: BoxDecoration(
@@ -70,23 +76,26 @@ class GenieScreen extends StatelessWidget {
                         child: Column(
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: <Widget>[
-                            _HeaderView(
+                            const _HeaderView(
                               title: 'Pickup or Drop any items',
                               buttonTitle: 'ADD PICKUP DROP DETAILS',
                             ),
-                            CustomDividerView(dividerHeight: 3.0),
-                            UIHelper.verticalSpaceMedium(),
+                            const CustomDividerView(dividerHeight: 3.0),
+                            UIHelper.verticalSpaceMedium,
                             Text(
                               'Some things we can pick or drop for you',
-                              style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 14.0),
+                              style: Theme.of(context)
+                                  .textTheme
+                                  .bodyText1!
+                                  .copyWith(fontSize: 14.0),
                             ),
-                            UIHelper.verticalSpaceMedium(),
+                            UIHelper.verticalSpaceMedium,
                             LimitedBox(
                               maxHeight: 100.0,
                               child: ListView.builder(
                                 itemCount: services.length,
                                 scrollDirection: Axis.horizontal,
-                                itemBuilder: (context, index) => Container(
+                                itemBuilder: (context, index) => SizedBox(
                                   width: 80.0,
                                   child: Column(
                                     mainAxisSize: MainAxisSize.min,
@@ -96,7 +105,7 @@ class GenieScreen extends StatelessWidget {
                                           padding: const EdgeInsets.all(10.0),
                                           decoration: BoxDecoration(
                                             color: Colors.grey[200],
-                                            boxShadow: <BoxShadow>[
+                                            boxShadow: const <BoxShadow>[
                                               BoxShadow(
                                                 color: Colors.grey,
                                                 blurRadius: 3.0,
@@ -116,7 +125,10 @@ class GenieScreen extends StatelessWidget {
                                         services[index].title,
                                         textAlign: TextAlign.center,
                                         maxLines: 2,
-                                        style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 13.5),
+                                        style: Theme.of(context)
+                                            .textTheme
+                                            .bodyText1!
+                                            .copyWith(fontSize: 13.5),
                                       )
                                     ],
                                   ),
@@ -126,14 +138,14 @@ class GenieScreen extends StatelessWidget {
                           ],
                         ),
                       ),
-                      UIHelper.verticalSpaceMedium(),
+                      UIHelper.verticalSpaceMedium,
                       Container(
                         padding: const EdgeInsets.all(20.0),
                         decoration: BoxDecoration(
                           color: Colors.white,
                           borderRadius: BorderRadius.circular(10.0),
                         ),
-                        child: _HeaderView(
+                        child: const _HeaderView(
                           title: 'Buy Anything from any store',
                           buttonTitle: 'FIND A STORE',
                         ),
@@ -155,9 +167,9 @@ class _HeaderView extends StatelessWidget {
   final String buttonTitle;
 
   const _HeaderView({
-    Key key,
-    @required this.title,
-    @required this.buttonTitle,
+    Key? key,
+    required this.title,
+    required this.buttonTitle,
   }) : super(key: key);
 
   @override
@@ -167,12 +179,12 @@ class _HeaderView extends StatelessWidget {
       children: <Widget>[
         Text(
           title,
-          style: Theme.of(context).textTheme.headline6.copyWith(
+          style: Theme.of(context).textTheme.headline6!.copyWith(
                 fontSize: 17.0,
                 fontWeight: FontWeight.bold,
               ),
         ),
-        UIHelper.verticalSpaceMedium(),
+        UIHelper.verticalSpaceMedium,
         Container(
           padding: const EdgeInsets.symmetric(horizontal: 15.0),
           height: 50.0,
@@ -181,12 +193,15 @@ class _HeaderView extends StatelessWidget {
             color: darkOrange,
             child: Text(
               buttonTitle,
-              style: Theme.of(context).textTheme.subtitle2.copyWith(color: Colors.white, fontSize: 14.0),
+              style: Theme.of(context)
+                  .textTheme
+                  .subtitle2!
+                  .copyWith(color: Colors.white, fontSize: 14.0),
             ),
             onPressed: () {},
           ),
         ),
-        UIHelper.verticalSpaceMedium(),
+        UIHelper.verticalSpaceMedium,
       ],
     );
   }
