@@ -4,6 +4,8 @@ import 'package:swiggy_ui/utils/ui_helper.dart';
 import 'package:swiggy_ui/widgets/dotted_seperator_view.dart';
 
 class GenieView extends StatelessWidget {
+  const GenieView({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -19,15 +21,18 @@ class GenieView extends StatelessWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: <Widget>[
                     Text('Genie', style: Theme.of(context).textTheme.headline4),
-                    UIHelper.verticalSpaceSmall(),
+                    UIHelper.verticalSpaceSmall,
                     Text(
                       'Anything you need, deliverd.\nPick-up, Drop or Buy anything,\nfrom anywhere in your city',
-                      style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey),
+                      style: Theme.of(context)
+                          .textTheme
+                          .bodyText1!
+                          .copyWith(color: Colors.grey),
                     )
                   ],
                 ),
               ),
-              UIHelper.horizontalSpaceMedium(),
+              UIHelper.horizontalSpaceMedium,
               LimitedBox(
                 maxWidth: 100.0,
                 child: Image.asset(
@@ -37,13 +42,13 @@ class GenieView extends StatelessWidget {
               ),
             ],
           ),
-          UIHelper.verticalSpaceMedium(),
-          DottedSeperatorView(),
-          UIHelper.verticalSpaceMedium(),
+          UIHelper.verticalSpaceMedium,
+          const DottedSeperatorView(),
+          UIHelper.verticalSpaceMedium,
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
+            children: const <Widget>[
               _GenieCardView(
                 title: 'Buy\nAnything',
                 desc: 'Stationery\nMedicine\nGrocery\n& more',
@@ -64,17 +69,17 @@ class GenieView extends StatelessWidget {
 
 class _GenieCardView extends StatelessWidget {
   const _GenieCardView({
-    Key key,
-    @required this.title,
-    @required this.desc,
-    @required this.image,
+    Key? key,
+    required this.title,
+    required this.desc,
+    required this.image,
     this.onTap,
   }) : super(key: key);
 
   final String title;
   final String desc;
   final String image;
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   @override
   Widget build(BuildContext context) {
@@ -89,9 +94,9 @@ class _GenieCardView extends StatelessWidget {
             color: Colors.white,
             boxShadow: <BoxShadow>[
               BoxShadow(
-                color: Colors.grey[200],
+                color: Colors.grey[200]!,
                 blurRadius: 2.0,
-                offset: Offset(1.0, 3.0),
+                offset: const Offset(1.0, 3.0),
               )
             ],
           ),
@@ -100,9 +105,12 @@ class _GenieCardView extends StatelessWidget {
             children: <Widget>[
               Text(
                 title,
-                style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 22.0),
+                style: Theme.of(context)
+                    .textTheme
+                    .subtitle2!
+                    .copyWith(fontSize: 22.0),
               ),
-              UIHelper.verticalSpaceMedium(),
+              UIHelper.verticalSpaceMedium,
               Row(
                 crossAxisAlignment: CrossAxisAlignment.end,
                 children: <Widget>[
@@ -113,31 +121,31 @@ class _GenieCardView extends StatelessWidget {
                         desc,
                         style: Theme.of(context).textTheme.bodyText1,
                       ),
-                      UIHelper.verticalSpaceSmall(),
+                      UIHelper.verticalSpaceSmall,
                       ClipOval(
                         child: Container(
                           alignment: Alignment.center,
                           color: swiggyOrange,
                           height: 25.0,
                           width: 25.0,
-                          child: Icon(
+                          child: const Icon(
                             Icons.arrow_forward_ios,
                             size: 12.0,
                             color: Colors.white,
                           ),
                         ),
                       ),
-                      UIHelper.verticalSpaceMedium(),
+                      UIHelper.verticalSpaceMedium,
                     ],
                   ),
-                  UIHelper.horizontalSpaceMedium(),
+                  UIHelper.horizontalSpaceMedium,
                   Flexible(
                     child: Image.asset(
                       image,
                       fit: BoxFit.contain,
                     ),
                   ),
-                  UIHelper.horizontalSpaceExtraSmall(),
+                  UIHelper.horizontalSpaceExtraSmall,
                 ],
               )
             ],
