@@ -5,16 +5,18 @@ import 'package:swiggy_ui/widgets/custom_divider_view.dart';
 import 'package:swiggy_ui/widgets/veg_badge_view.dart';
 
 class RestaurantDetailScreen extends StatelessWidget {
+  const RestaurantDetailScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         elevation: 0.0,
-        actions: <Widget>[
+        actions: const <Widget>[
           Icon(Icons.favorite_border),
-          UIHelper.horizontalSpaceSmall(),
+          UIHelper.horizontalSpaceSmall,
           Icon(Icons.search),
-          UIHelper.horizontalSpaceSmall(),
+          UIHelper.horizontalSpaceSmall,
         ],
       ),
       body: _OrderNowView(),
@@ -25,97 +27,106 @@ class RestaurantDetailScreen extends StatelessWidget {
 class _OrderNowView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.all(15.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text(
-                    'Namma Veedu Vasanta Bhavan',
-                    style: Theme.of(context).textTheme.subtitle2.copyWith(fontWeight: FontWeight.bold, fontSize: 16.0),
-                  ),
-                  UIHelper.verticalSpaceSmall(),
-                  Text('South Indian', style: Theme.of(context).textTheme.bodyText1),
-                  UIHelper.verticalSpaceExtraSmall(),
-                  Text('Velachery Main Road, Madipakkam', style: Theme.of(context).textTheme.bodyText1),
-                  UIHelper.verticalSpaceMedium(),
-                  CustomDividerView(dividerHeight: 1.0),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+    return SingleChildScrollView(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.all(15.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Text(
+                  'Namma Veedu Vasanta Bhavan',
+                  style: Theme.of(context)
+                      .textTheme
+                      .subtitle2!
+                      .copyWith(fontWeight: FontWeight.bold, fontSize: 16.0),
+                ),
+                UIHelper.verticalSpaceSmall,
+                Text('South Indian',
+                    style: Theme.of(context).textTheme.bodyText1),
+                UIHelper.verticalSpaceExtraSmall,
+                Text('Velachery Main Road, Madipakkam',
+                    style: Theme.of(context).textTheme.bodyText1),
+                UIHelper.verticalSpaceMedium,
+                const CustomDividerView(dividerHeight: 1.0),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: <Widget>[
+                    _buildVerticalStack(context, '4.1', 'Packaging 80%'),
+                    _buildVerticalStack(context, '29 mins', 'Delivery Time'),
+                    _buildVerticalStack(context, 'Rs150', 'For Two'),
+                  ],
+                ),
+                const CustomDividerView(dividerHeight: 1.0),
+                UIHelper.verticalSpaceMedium,
+                Column(
+                  children: <Widget>[
+                    _buildOfferTile(
+                        context, '30% off up to Rs75 | Use code SWIGGYIT'),
+                    _buildOfferTile(context,
+                        '20% off up to Rs100 with SBI credit cards, once per week | Use code 100SBI')
+                  ],
+                ),
+                UIHelper.verticalSpaceSmall,
+              ],
+            ),
+          ),
+          const CustomDividerView(dividerHeight: 15.0),
+          Container(
+            height: 80.0,
+            padding: const EdgeInsets.all(10.0),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                Expanded(
+                  child: Row(
+                    crossAxisAlignment: CrossAxisAlignment.center,
                     children: <Widget>[
-                      _buildVerticalStack(context, '4.1', 'Packaging 80%'),
-                      _buildVerticalStack(context, '29 mins', 'Delivery Time'),
-                      _buildVerticalStack(context, 'Rs150', 'For Two'),
+                      const Icon(Icons.filter_vintage,
+                          color: Colors.green, size: 12.0),
+                      UIHelper.horizontalSpaceExtraSmall,
+                      Text('PURE VEG',
+                          style: Theme.of(context)
+                              .textTheme
+                              .subtitle2!
+                              .copyWith(
+                                  fontWeight: FontWeight.bold, fontSize: 16.0))
                     ],
                   ),
-                  CustomDividerView(dividerHeight: 1.0),
-                  UIHelper.verticalSpaceMedium(),
-                  Column(
-                    children: <Widget>[
-                      _buildOfferTile(context, '30% off up to Rs75 | Use code SWIGGYIT'),
-                      _buildOfferTile(
-                          context, '20% off up to Rs100 with SBI credit cards, once per week | Use code 100SBI')
-                    ],
-                  ),
-                  UIHelper.verticalSpaceSmall(),
-                ],
-              ),
+                ),
+                const CustomDividerView(dividerHeight: 0.5, color: Colors.black)
+              ],
             ),
-            CustomDividerView(dividerHeight: 15.0),
-            Container(
-              height: 80.0,
-              padding: const EdgeInsets.all(10.0),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Expanded(
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: <Widget>[
-                        Icon(Icons.filter_vintage, color: Colors.green, size: 12.0),
-                        UIHelper.horizontalSpaceExtraSmall(),
-                        Text('PURE VEG',
-                            style: Theme.of(context)
-                                .textTheme
-                                .subtitle2
-                                .copyWith(fontWeight: FontWeight.bold, fontSize: 16.0))
-                      ],
-                    ),
-                  ),
-                  CustomDividerView(dividerHeight: 0.5, color: Colors.black)
-                ],
-              ),
+          ),
+          Padding(
+            padding: const EdgeInsets.all(10.0),
+            child: Text(
+              'Recommended',
+              style: Theme.of(context)
+                  .textTheme
+                  .subtitle2!
+                  .copyWith(fontSize: 18.0),
             ),
-            Padding(
-              padding: const EdgeInsets.all(10.0),
-              child: Text(
-                'Recommended',
-                style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 18.0),
-              ),
-            ),
-            _RecommendedFoodView(),
-            CustomDividerView(dividerHeight: 15.0),
-            _FoodListView(
-              title: 'Breakfast',
-              foods: RestaurantDetail.getBreakfast(),
-            ),
-            CustomDividerView(dividerHeight: 15.0),
-            _FoodListView(
-              title: 'All Time Favourite',
-              foods: RestaurantDetail.getAllTimeFavFoods(),
-            ),
-            CustomDividerView(dividerHeight: 15.0),
-            _FoodListView(
-              title: 'Kozhukattaiyum & Paniyarams',
-              foods: RestaurantDetail.getOtherDishes(),
-            )
-          ],
-        ),
+          ),
+          _RecommendedFoodView(),
+          const CustomDividerView(dividerHeight: 15.0),
+          _FoodListView(
+            title: 'Breakfast',
+            foods: RestaurantDetail.getBreakfast(),
+          ),
+          const CustomDividerView(dividerHeight: 15.0),
+          _FoodListView(
+            title: 'All Time Favourite',
+            foods: RestaurantDetail.getAllTimeFavFoods(),
+          ),
+          const CustomDividerView(dividerHeight: 15.0),
+          _FoodListView(
+            title: 'Kozhukattaiyum & Paniyarams',
+            foods: RestaurantDetail.getOtherDishes(),
+          )
+        ],
       ),
     );
   }
@@ -125,18 +136,23 @@ class _OrderNowView extends StatelessWidget {
         child: Row(
           children: <Widget>[
             Icon(Icons.local_offer, color: Colors.red[600], size: 15.0),
-            UIHelper.horizontalSpaceSmall(),
+            UIHelper.horizontalSpaceSmall,
             Flexible(
               child: Text(
                 desc,
-                style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 13.0),
+                style: Theme.of(context)
+                    .textTheme
+                    .bodyText1!
+                    .copyWith(fontSize: 13.0),
               ),
             )
           ],
         ),
       );
 
-  Expanded _buildVerticalStack(BuildContext context, String title, String subtitle) => Expanded(
+  Expanded _buildVerticalStack(
+          BuildContext context, String title, String subtitle) =>
+      Expanded(
         child: SizedBox(
           height: 60.0,
           child: Column(
@@ -145,10 +161,17 @@ class _OrderNowView extends StatelessWidget {
             children: <Widget>[
               Text(
                 title,
-                style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 15.0),
+                style: Theme.of(context)
+                    .textTheme
+                    .subtitle2!
+                    .copyWith(fontSize: 15.0),
               ),
-              UIHelper.verticalSpaceExtraSmall(),
-              Text(subtitle, style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 13.0))
+              UIHelper.verticalSpaceExtraSmall,
+              Text(subtitle,
+                  style: Theme.of(context)
+                      .textTheme
+                      .bodyText1!
+                      .copyWith(fontSize: 13.0))
             ],
           ),
         ),
@@ -166,7 +189,7 @@ class _RecommendedFoodView extends StatelessWidget {
         shrinkWrap: true,
         crossAxisCount: 2,
         childAspectRatio: 0.8,
-        physics: NeverScrollableScrollPhysics(),
+        physics: const NeverScrollableScrollPhysics(),
         children: List.generate(
           foods.length,
           (index) => Container(
@@ -180,7 +203,7 @@ class _RecommendedFoodView extends StatelessWidget {
                     fit: BoxFit.fill,
                   ),
                 ),
-                UIHelper.verticalSpaceExtraSmall(),
+                UIHelper.verticalSpaceExtraSmall,
                 SizedBox(
                   height: 80.0,
                   child: Column(
@@ -189,32 +212,36 @@ class _RecommendedFoodView extends StatelessWidget {
                     children: <Widget>[
                       Text(
                         'BREAKFAST',
-                        style: Theme.of(context).textTheme.bodyText1.copyWith(
+                        style: Theme.of(context).textTheme.bodyText1!.copyWith(
                               fontSize: 10.0,
                               color: Colors.grey[700],
                             ),
                       ),
-                      UIHelper.verticalSpaceExtraSmall(),
+                      UIHelper.verticalSpaceExtraSmall,
                       Row(
                         children: <Widget>[
-                          VegBadgeView(),
-                          UIHelper.horizontalSpaceExtraSmall(),
+                          const VegBadgeView(),
+                          UIHelper.horizontalSpaceExtraSmall,
                           Flexible(
                             child: Text(
                               foods[index].title,
                               maxLines: 1,
-                              style: TextStyle(fontWeight: FontWeight.bold),
+                              style:
+                                  const TextStyle(fontWeight: FontWeight.bold),
                             ),
                           ),
                         ],
                       ),
-                      UIHelper.verticalSpaceMedium(),
+                      UIHelper.verticalSpaceMedium,
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: <Widget>[
                           Text(foods[index].price,
-                              style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 14.0)),
-                          AddBtnView()
+                              style: Theme.of(context)
+                                  .textTheme
+                                  .bodyText1!
+                                  .copyWith(fontSize: 14.0)),
+                          const AddBtnView()
                         ],
                       )
                     ],
@@ -231,7 +258,7 @@ class _RecommendedFoodView extends StatelessWidget {
 
 class AddBtnView extends StatelessWidget {
   const AddBtnView({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -243,7 +270,10 @@ class AddBtnView extends StatelessWidget {
       ),
       child: Text(
         'ADD',
-        style: Theme.of(context).textTheme.subtitle2.copyWith(color: Colors.green),
+        style: Theme.of(context)
+            .textTheme
+            .subtitle2!
+            .copyWith(color: Colors.green),
       ),
     );
   }
@@ -251,9 +281,9 @@ class AddBtnView extends StatelessWidget {
 
 class _FoodListView extends StatelessWidget {
   const _FoodListView({
-    Key key,
-    @required this.title,
-    @required this.foods,
+    Key? key,
+    required this.title,
+    required this.foods,
   }) : super(key: key);
 
   final String title;
@@ -267,26 +297,27 @@ class _FoodListView extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
-          UIHelper.verticalSpaceMedium(),
+          UIHelper.verticalSpaceMedium,
           Text(
             title,
-            style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 18.0),
+            style:
+                Theme.of(context).textTheme.subtitle2!.copyWith(fontSize: 18.0),
           ),
           ListView.builder(
             shrinkWrap: true,
             itemCount: foods.length,
-            physics: NeverScrollableScrollPhysics(),
+            physics: const NeverScrollableScrollPhysics(),
             itemBuilder: (context, index) => Container(
               padding: const EdgeInsets.symmetric(vertical: 10.0),
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
-                  UIHelper.verticalSpaceSmall(),
+                  UIHelper.verticalSpaceSmall,
                   Row(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: <Widget>[
-                      VegBadgeView(),
-                      UIHelper.horizontalSpaceMedium(),
+                      const VegBadgeView(),
+                      UIHelper.horizontalSpaceMedium,
                       Expanded(
                         child: Column(
                           crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -295,18 +326,24 @@ class _FoodListView extends StatelessWidget {
                               foods[index].title,
                               style: Theme.of(context).textTheme.bodyText1,
                             ),
-                            UIHelper.verticalSpaceSmall(),
+                            UIHelper.verticalSpaceSmall,
                             Text(
                               foods[index].price,
-                              style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 14.0),
+                              style: Theme.of(context)
+                                  .textTheme
+                                  .bodyText1!
+                                  .copyWith(fontSize: 14.0),
                             ),
-                            UIHelper.verticalSpaceMedium(),
+                            UIHelper.verticalSpaceMedium,
                             if (foods[index].desc != null)
                               Text(
-                                foods[index].desc,
+                                foods[index].desc!,
                                 maxLines: 2,
                                 overflow: TextOverflow.ellipsis,
-                                style: Theme.of(context).textTheme.bodyText1.copyWith(
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .bodyText1!
+                                    .copyWith(
                                       fontSize: 12.0,
                                       color: Colors.grey[500],
                                     ),
@@ -314,7 +351,7 @@ class _FoodListView extends StatelessWidget {
                           ],
                         ),
                       ),
-                      AddBtnView()
+                      const AddBtnView()
                     ],
                   ),
                 ],
