@@ -12,9 +12,9 @@ class RestaurantVerticalListView extends StatelessWidget {
   final bool isAllRestaurantNearby;
 
   const RestaurantVerticalListView({
-    Key key,
-    @required this.title,
-    @required this.restaurants,
+    Key? key,
+    required this.title,
+    required this.restaurants,
     this.isAllRestaurantNearby = false,
   })  : assert(title != ''),
         super(key: key);
@@ -31,24 +31,28 @@ class RestaurantVerticalListView extends StatelessWidget {
         children: <Widget>[
           Text(
             title,
-            style: Theme.of(context).textTheme.headline4.copyWith(fontSize: 20.0),
+            style:
+                Theme.of(context).textTheme.headline4!.copyWith(fontSize: 20.0),
           ),
           isAllRestaurantNearby
               ? Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: <Widget>[
-                    UIHelper.verticalSpaceExtraSmall(),
+                    UIHelper.verticalSpaceExtraSmall,
                     Text(
                       'Discover unique tastes near you',
-                      style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 14.0),
+                      style: Theme.of(context)
+                          .textTheme
+                          .bodyText1!
+                          .copyWith(fontSize: 14.0),
                     ),
                   ],
                 )
-              : SizedBox(),
-          UIHelper.verticalSpaceMedium(),
+              : const SizedBox(),
+          UIHelper.verticalSpaceMedium,
           ListView.builder(
             shrinkWrap: true,
-            physics: NeverScrollableScrollPhysics(),
+            physics: const NeverScrollableScrollPhysics(),
             itemCount: restaurants.length,
             itemBuilder: (context, index) => InkWell(
               onTap: isTabletDesktop
@@ -57,7 +61,7 @@ class RestaurantVerticalListView extends StatelessWidget {
                       Navigator.push(
                         context,
                         MaterialPageRoute(
-                          builder: (context) => RestaurantDetailScreen(),
+                          builder: (context) => const RestaurantDetailScreen(),
                         ),
                       );
                     },
