@@ -4,7 +4,7 @@ import 'package:swiggy_ui/utils/app_colors.dart';
 import 'package:swiggy_ui/utils/ui_helper.dart';
 
 class CartView extends StatelessWidget {
-  const CartView({Key key, this.isTab = false}) : super(key: key);
+  const CartView({Key? key, this.isTab = false}) : super(key: key);
 
   final bool isTab;
 
@@ -15,7 +15,11 @@ class CartView extends StatelessWidget {
       child: Container(
         color: Colors.white,
         alignment: Alignment.center,
-        padding: EdgeInsets.only(left: isTab ? 20.0 : 40.0, top: 40.0, right: isTab ? 20.0 : 40.0, bottom: 20.0),
+        padding: EdgeInsets.only(
+            left: isTab ? 20.0 : 40.0,
+            top: 40.0,
+            right: isTab ? 20.0 : 40.0,
+            bottom: 20.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
@@ -54,11 +58,11 @@ class _UserHeader extends StatelessWidget {
                     ),
                     height: 50.0,
                     width: 50.0,
-                    child: Icon(
+                    child: const Icon(
                       Icons.notifications_outlined,
                     ),
                   ),
-                  UIHelper.horizontalSpaceMedium(),
+                  UIHelper.horizontalSpaceMedium,
                   ClipOval(
                     child: Image.asset(
                       'assets/images/user.jpg',
@@ -67,25 +71,28 @@ class _UserHeader extends StatelessWidget {
                       fit: BoxFit.fill,
                     ),
                   ),
-                  UIHelper.horizontalSpaceMedium(),
+                  UIHelper.horizontalSpaceMedium,
                   Column(
                     mainAxisAlignment: MainAxisAlignment.center,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       Text(
                         'Vinoth',
-                        style:
-                            Theme.of(context).textTheme.headline6.copyWith(fontSize: 17.0, fontWeight: FontWeight.bold),
+                        style: Theme.of(context).textTheme.headline6!.copyWith(
+                            fontSize: 17.0, fontWeight: FontWeight.bold),
                       ),
-                      UIHelper.verticalSpaceExtraSmall(),
+                      UIHelper.verticalSpaceExtraSmall,
                       Text(
                         'User',
-                        style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey, fontSize: 13.0),
+                        style: Theme.of(context)
+                            .textTheme
+                            .bodyText1!
+                            .copyWith(color: Colors.grey, fontSize: 13.0),
                       ),
                     ],
                   ),
-                  UIHelper.horizontalSpaceMedium(),
-                  Icon(Icons.keyboard_arrow_down_outlined),
+                  UIHelper.horizontalSpaceMedium,
+                  const Icon(Icons.keyboard_arrow_down_outlined),
                 ],
               ),
             ),
@@ -110,24 +117,31 @@ class _MyOrdersList extends StatelessWidget {
           children: [
             Row(
               children: [
-                Text('Order Menu', style: Theme.of(context).textTheme.headline6),
-                Spacer(),
-                Text('See all', style: Theme.of(context).textTheme.subtitle1.copyWith(color: swiggyOrange)),
+                Text('Order Menu',
+                    style: Theme.of(context).textTheme.headline6),
+                const Spacer(),
+                Text('See all',
+                    style: Theme.of(context)
+                        .textTheme
+                        .subtitle1!
+                        .copyWith(color: swiggyOrange)),
               ],
             ),
-            UIHelper.verticalSpaceSmall(),
+            UIHelper.verticalSpaceSmall,
             Expanded(
               child: ListView(
                 shrinkWrap: true,
                 children: List.generate(
                   cartItems.length,
                   (index) => Container(
-                    margin: const EdgeInsets.symmetric(vertical: 5.0, horizontal: 4.0),
-                    padding: const EdgeInsets.symmetric(vertical: 15.0, horizontal: 10.0),
+                    margin: const EdgeInsets.symmetric(
+                        vertical: 5.0, horizontal: 4.0),
+                    padding: const EdgeInsets.symmetric(
+                        vertical: 15.0, horizontal: 10.0),
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(18.0),
                       color: Colors.white,
-                      boxShadow: [
+                      boxShadow: const [
                         BoxShadow(color: Colors.grey),
                       ],
                     ),
@@ -143,30 +157,37 @@ class _MyOrdersList extends StatelessWidget {
                             fit: BoxFit.fill,
                           ),
                         ),
-                        UIHelper.horizontalSpaceSmall(),
+                        UIHelper.horizontalSpaceSmall,
                         Expanded(
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
                               Text(cartItems[index].title,
-                                  style: Theme.of(context).textTheme.headline6.copyWith(fontSize: 14.0)),
-                              UIHelper.verticalSpaceMedium(),
+                                  style: Theme.of(context)
+                                      .textTheme
+                                      .headline6!
+                                      .copyWith(fontSize: 14.0)),
+                              UIHelper.verticalSpaceMedium,
                               Row(
                                 children: [
-                                  Icon(Icons.close, size: 18.0),
-                                  UIHelper.horizontalSpaceMedium(),
+                                  const Icon(Icons.close, size: 18.0),
+                                  UIHelper.horizontalSpaceMedium,
                                   Container(
-                                    padding: const EdgeInsets.symmetric(vertical: 2.0, horizontal: 10.0),
+                                    padding: const EdgeInsets.symmetric(
+                                        vertical: 2.0, horizontal: 10.0),
                                     decoration: BoxDecoration(
-                                      border: Border.all(color: Colors.grey[300]),
+                                      border:
+                                          Border.all(color: Colors.grey[300]!),
                                       borderRadius: BorderRadius.circular(8.0),
                                       color: Colors.grey[100],
                                     ),
                                     child: Row(
-                                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceBetween,
                                       children: [
                                         Text('${cartItems[index].quantity}'),
-                                        Icon(Icons.keyboard_arrow_down_outlined),
+                                        const Icon(
+                                            Icons.keyboard_arrow_down_outlined),
                                       ],
                                     ),
                                   ),
@@ -177,10 +198,8 @@ class _MyOrdersList extends StatelessWidget {
                         ),
                         Text(
                           'Rs ${cartItems[index].price}',
-                          style: Theme.of(context)
-                              .textTheme
-                              .headline6
-                              .copyWith(fontSize: 16.0, fontWeight: FontWeight.w800),
+                          style: Theme.of(context).textTheme.headline6!.copyWith(
+                              fontSize: 16.0, fontWeight: FontWeight.w800),
                         ),
                       ],
                     ),
@@ -198,8 +217,14 @@ class _MyOrdersList extends StatelessWidget {
 class _Checkout extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final listTileStyle = Theme.of(context).textTheme.subtitle1.copyWith(fontSize: 14.0, fontWeight: FontWeight.w600);
-    final amountStyle = Theme.of(context).textTheme.headline6.copyWith(fontSize: 15.0, fontWeight: FontWeight.bold);
+    final listTileStyle = Theme.of(context)
+        .textTheme
+        .subtitle1!
+        .copyWith(fontSize: 14.0, fontWeight: FontWeight.w600);
+    final amountStyle = Theme.of(context)
+        .textTheme
+        .headline6!
+        .copyWith(fontSize: 15.0, fontWeight: FontWeight.bold);
 
     return Expanded(
       flex: 3,
@@ -219,21 +244,22 @@ class _Checkout extends StatelessWidget {
           ),
           Container(
             margin: const EdgeInsets.symmetric(vertical: 15.0),
-            padding: const EdgeInsets.symmetric(vertical: 10.0, horizontal: 10.0),
+            padding:
+                const EdgeInsets.symmetric(vertical: 10.0, horizontal: 10.0),
             decoration: BoxDecoration(
               border: Border.all(color: swiggyOrange),
               color: Colors.deepOrange[50],
             ),
             child: Row(
               children: [
-                Expanded(child: Text('Find Promotion')),
-                UIHelper.horizontalSpaceMedium(),
+                const Expanded(child: Text('Find Promotion')),
+                UIHelper.horizontalSpaceMedium,
                 SizedBox(
                   height: 38.0,
                   child: ElevatedButton.icon(
                     onPressed: () {},
-                    icon: Icon(Icons.countertops_outlined),
-                    label: Text('Add Coupon'),
+                    icon: const Icon(Icons.countertops_outlined),
+                    label: const Text('Add Coupon'),
                     style: ElevatedButton.styleFrom(
                       onPrimary: Colors.white,
                       primary: swiggyOrange,
@@ -250,14 +276,14 @@ class _Checkout extends StatelessWidget {
             title: Text('To Pay', style: listTileStyle),
             trailing: Text('Rs 540', style: amountStyle),
           ),
-          UIHelper.verticalSpaceMedium(),
+          UIHelper.verticalSpaceMedium,
           SizedBox(
             height: 45.0,
             width: double.infinity,
             child: ElevatedButton.icon(
               onPressed: () {},
-              icon: Icon(Icons.countertops_outlined),
-              label: Text('Checkout'),
+              icon: const Icon(Icons.countertops_outlined),
+              label: const Text('Checkout'),
               style: ElevatedButton.styleFrom(
                 onPrimary: Colors.white,
                 primary: swiggyOrange,
