@@ -4,7 +4,8 @@ import 'package:swiggy_ui/utils/app_colors.dart';
 import 'package:swiggy_ui/utils/ui_helper.dart';
 
 class MenuView extends StatelessWidget {
-  const MenuView({Key key, this.expandFlex = 2, this.isTab = false}) : super(key: key);
+  const MenuView({Key? key, this.expandFlex = 2, this.isTab = false})
+      : super(key: key);
 
   final int expandFlex;
   final bool isTab;
@@ -18,7 +19,11 @@ class MenuView extends StatelessWidget {
       child: Container(
         color: Colors.white,
         alignment: Alignment.center,
-        padding: EdgeInsets.only(left: isTab ? 20.0 : 40.0, top: 40.0, right: isTab ? 20.0 : 40.0, bottom: 20.0),
+        padding: EdgeInsets.only(
+            left: isTab ? 20.0 : 40.0,
+            top: 40.0,
+            right: isTab ? 20.0 : 40.0,
+            bottom: 20.0),
         child: Column(
           children: [
             ListView(
@@ -31,10 +36,10 @@ class MenuView extends StatelessWidget {
                 ),
               ),
             ),
-            Spacer(),
+            const Spacer(),
             isTab
                 ? IconButton(
-                    icon: Icon(Icons.exit_to_app),
+                    icon: const Icon(Icons.exit_to_app),
                     iconSize: 30.0,
                     onPressed: () {},
                   )
@@ -43,8 +48,8 @@ class MenuView extends StatelessWidget {
                     child: SizedBox(
                       height: 52.0,
                       child: OutlinedButton.icon(
-                        icon: Icon(Icons.exit_to_app),
-                        label: Text('Logout'),
+                        icon: const Icon(Icons.exit_to_app),
+                        label: const Text('Logout'),
                         onPressed: () => print("it's pressed"),
                         style: ElevatedButton.styleFrom(
                           onPrimary: swiggyOrange,
@@ -64,7 +69,8 @@ class MenuView extends StatelessWidget {
 }
 
 class _MenuItem extends StatefulWidget {
-  const _MenuItem({Key key, @required this.menu, this.isTab = false}) : super(key: key);
+  const _MenuItem({Key? key, required this.menu, this.isTab = false})
+      : super(key: key);
 
   final Menu menu;
   final bool isTab;
@@ -81,8 +87,10 @@ class __MenuItemState extends State<_MenuItem> {
   Widget build(BuildContext context) {
     return Container(
       alignment: Alignment.center,
-      margin: isTab ? EdgeInsets.zero : const EdgeInsets.symmetric(vertical: 5.0),
-      padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: isTab ? 0.0 : 10.0),
+      margin:
+          isTab ? EdgeInsets.zero : const EdgeInsets.symmetric(vertical: 5.0),
+      padding:
+          EdgeInsets.symmetric(vertical: 5.0, horizontal: isTab ? 0.0 : 10.0),
       child: InkWell(
         onTap: () {
           print('Tapped');
@@ -103,10 +111,12 @@ class __MenuItemState extends State<_MenuItem> {
               : null,
           padding: isTab
               ? const EdgeInsets.symmetric(vertical: 10.0)
-              : const EdgeInsets.only(left: 15.0, top: 10.0, right: 25.0, bottom: 10.0),
+              : const EdgeInsets.only(
+                  left: 15.0, top: 10.0, right: 25.0, bottom: 10.0),
           child: isTab
               ? IconButton(
-                  icon: Icon(widget.menu.icon, color: isHovered ? swiggyOrange : Colors.black),
+                  icon: Icon(widget.menu.icon,
+                      color: isHovered ? swiggyOrange : Colors.black),
                   iconSize: 30.0,
                   onPressed: () {},
                 )
@@ -115,11 +125,13 @@ class __MenuItemState extends State<_MenuItem> {
                   mainAxisAlignment: MainAxisAlignment.start,
                   crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
-                    Icon(widget.menu.icon, color: isHovered ? swiggyOrange : Colors.black, size: 30.0),
-                    UIHelper.horizontalSpaceMedium(),
+                    Icon(widget.menu.icon,
+                        color: isHovered ? swiggyOrange : Colors.black,
+                        size: 30.0),
+                    UIHelper.horizontalSpaceMedium,
                     Text(
                       widget.menu.title,
-                      style: Theme.of(context).textTheme.headline6.copyWith(
+                      style: Theme.of(context).textTheme.headline6!.copyWith(
                             color: isHovered ? swiggyOrange : Colors.black,
                           ),
                     ),
