@@ -7,8 +7,8 @@ class FoodListItemView extends StatelessWidget {
   final SpotlightBestTopFood restaurant;
 
   const FoodListItemView({
-    Key key,
-    @required this.restaurant,
+    Key? key,
+    required this.restaurant,
   }) : super(key: key);
 
   @override
@@ -21,7 +21,7 @@ class FoodListItemView extends StatelessWidget {
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10.0),
               color: Colors.white,
-              boxShadow: <BoxShadow>[
+              boxShadow: const <BoxShadow>[
                 BoxShadow(
                   color: Colors.grey,
                   blurRadius: 2.0,
@@ -35,23 +35,32 @@ class FoodListItemView extends StatelessWidget {
               fit: BoxFit.fill,
             ),
           ),
-          UIHelper.horizontalSpaceMedium(),
+          UIHelper.horizontalSpaceMedium,
           Column(
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               Text(
                 restaurant.name,
-                style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 18.0),
+                style: Theme.of(context)
+                    .textTheme
+                    .subtitle2!
+                    .copyWith(fontSize: 18.0),
               ),
               Text(restaurant.desc,
-                  style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey[800], fontSize: 13.5)),
-              UIHelper.verticalSpaceExtraSmall(),
+                  style: Theme.of(context)
+                      .textTheme
+                      .bodyText1!
+                      .copyWith(color: Colors.grey[800], fontSize: 13.5)),
+              UIHelper.verticalSpaceExtraSmall,
               Text(
                 restaurant.coupon,
-                style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.red[900], fontSize: 13.0),
+                style: Theme.of(context)
+                    .textTheme
+                    .bodyText1!
+                    .copyWith(color: Colors.red[900], fontSize: 13.0),
               ),
-              UIHelper.verticalSpaceSmall(),
+              UIHelper.verticalSpaceSmall,
               Row(
                 children: <Widget>[
                   Icon(
