@@ -4,8 +4,8 @@ import 'package:swiggy_ui/utils/ui_helper.dart';
 
 class SearchFoodListItemView extends StatelessWidget {
   const SearchFoodListItemView({
-    Key key,
-    @required this.food,
+    Key? key,
+    required this.food,
   }) : super(key: key);
 
   final SpotlightBestTopFood food;
@@ -21,7 +21,7 @@ class SearchFoodListItemView extends StatelessWidget {
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10.0),
               color: Colors.white,
-              boxShadow: <BoxShadow>[
+              boxShadow: const <BoxShadow>[
                 BoxShadow(
                   color: Colors.grey,
                   blurRadius: 2.0,
@@ -35,7 +35,7 @@ class SearchFoodListItemView extends StatelessWidget {
               fit: BoxFit.fill,
             ),
           ),
-          UIHelper.horizontalSpaceSmall(),
+          UIHelper.horizontalSpaceSmall,
           Expanded(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.start,
@@ -43,11 +43,17 @@ class SearchFoodListItemView extends StatelessWidget {
               children: <Widget>[
                 Text(
                   food.name,
-                  style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 15.0),
+                  style: Theme.of(context)
+                      .textTheme
+                      .subtitle2!
+                      .copyWith(fontSize: 15.0),
                 ),
                 Text(food.desc,
-                    style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey[600], fontSize: 13.5)),
-                UIHelper.verticalSpaceSmall(),
+                    style: Theme.of(context)
+                        .textTheme
+                        .bodyText1!
+                        .copyWith(color: Colors.grey[600], fontSize: 13.5)),
+                UIHelper.verticalSpaceSmall,
                 Row(
                   children: <Widget>[
                     Icon(
