@@ -7,12 +7,16 @@ import 'account/account_screen.dart';
 import 'cart/cart_screen.dart';
 
 class HomeBottomNavigationScreen extends StatefulWidget {
+  const HomeBottomNavigationScreen({Key? key}) : super(key: key);
+
   @override
-  _HomeBottomNavigationScreenState createState() => _HomeBottomNavigationScreenState();
+  _HomeBottomNavigationScreenState createState() =>
+      _HomeBottomNavigationScreenState();
 }
 
-class _HomeBottomNavigationScreenState extends State<HomeBottomNavigationScreen> {
-  final List<Widget> _children = [
+class _HomeBottomNavigationScreenState
+    extends State<HomeBottomNavigationScreen> {
+  final List<Widget> _children = const [
     SwiggyScreen(),
     SearchScreen(),
     CartScreen(),
@@ -23,7 +27,8 @@ class _HomeBottomNavigationScreenState extends State<HomeBottomNavigationScreen>
 
   @override
   Widget build(BuildContext context) {
-    final labelTextStyle = Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 8.0);
+    final labelTextStyle =
+        Theme.of(context).textTheme.subtitle2!.copyWith(fontSize: 8.0);
     return Scaffold(
       body: _children[selectedIndex],
       bottomNavigationBar: SizedBox(
@@ -40,7 +45,7 @@ class _HomeBottomNavigationScreenState extends State<HomeBottomNavigationScreen>
               selectedIndex = index;
             });
           },
-          items: [
+          items: const [
             BottomNavigationBarItem(
               icon: Icon(Icons.home),
               label: 'SWIGGY',
