@@ -2,8 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:swiggy_ui/views/mobile/home_bottom_navigation_screen.dart';
 
 class MobileScreen extends StatelessWidget {
+  const MobileScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return HomeBottomNavigationScreen();
+    return const HomeBottomNavigationScreen();
   }
 }
