@@ -6,12 +6,13 @@ import 'package:swiggy_ui/widgets/mobile/spotlight_best_top_food_item.dart';
 import 'package:swiggy_ui/widgets/responsive.dart';
 
 class BestInSafetyViews extends StatelessWidget {
-  final restaurants = SpotlightBestTopFood.getBestRestaurants();
+  const BestInSafetyViews({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final isTabletDesktop = Responsive.isTabletDesktop(context);
-    final customWidth = MediaQuery.of(context).size.width / (isTabletDesktop ? 3.8 : 1.1);
+    final customWidth =
+        MediaQuery.of(context).size.width / (isTabletDesktop ? 3.8 : 1.1);
 
     return Container(
       margin: const EdgeInsets.symmetric(vertical: 10.0),
@@ -25,27 +26,33 @@ class BestInSafetyViews extends StatelessWidget {
               children: <Widget>[
                 Row(
                   children: <Widget>[
-                    Icon(Icons.security),
-                    UIHelper.horizontalSpaceExtraSmall(),
+                    const Icon(Icons.security),
+                    UIHelper.horizontalSpaceExtraSmall,
                     Text(
                       'Best in Safety',
-                      style: Theme.of(context).textTheme.headline4.copyWith(fontSize: 20.0),
+                      style: Theme.of(context)
+                          .textTheme
+                          .headline4!
+                          .copyWith(fontSize: 20.0),
                     ),
-                    Spacer(),
+                    const Spacer(),
                     Row(
                       children: <Widget>[
                         Text(
                           'SEE ALL',
-                          style: Theme.of(context).textTheme.bodyText1.copyWith(fontWeight: FontWeight.bold),
+                          style: Theme.of(context)
+                              .textTheme
+                              .bodyText1!
+                              .copyWith(fontWeight: FontWeight.bold),
                         ),
-                        UIHelper.horizontalSpaceExtraSmall(),
+                        UIHelper.horizontalSpaceExtraSmall,
                         ClipOval(
                           child: Container(
                             alignment: Alignment.center,
                             color: swiggyOrange,
                             height: 25.0,
                             width: 25.0,
-                            child: Icon(
+                            child: const Icon(
                               Icons.arrow_forward_ios,
                               size: 12.0,
                               color: Colors.white,
@@ -56,28 +63,35 @@ class BestInSafetyViews extends StatelessWidget {
                     )
                   ],
                 ),
-                UIHelper.verticalSpaceExtraSmall(),
+                UIHelper.verticalSpaceExtraSmall,
                 Text(
                   'Restaurants with best safety standards',
-                  style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey),
+                  style: Theme.of(context)
+                      .textTheme
+                      .bodyText1!
+                      .copyWith(color: Colors.grey),
                 ),
               ],
             ),
           ),
-          UIHelper.verticalSpaceMedium(),
+          UIHelper.verticalSpaceMedium,
           LimitedBox(
             maxHeight: 270.0,
             child: ListView.builder(
               shrinkWrap: true,
               scrollDirection: Axis.horizontal,
-              itemCount: restaurants.length,
-              itemBuilder: (context, index) => Container(
+              itemCount: SpotlightBestTopFood.bestRestaurants.length,
+              itemBuilder: (context, index) => SizedBox(
                 width: customWidth,
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
-                    SpotlightBestTopFoodItem(restaurant: restaurants[index][0]),
-                    SpotlightBestTopFoodItem(restaurant: restaurants[index][1])
+                    SpotlightBestTopFoodItem(
+                        restaurant: SpotlightBestTopFood.bestRestaurants[index]
+                            [0]),
+                    SpotlightBestTopFoodItem(
+                        restaurant: SpotlightBestTopFood.bestRestaurants[index]
+                            [1])
                   ],
                 ),
               ),
