@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_swiper/flutter_swiper.dart';
+import 'package:card_swiper/card_swiper.dart';
 import 'package:swiggy_ui/widgets/responsive.dart';
 
 import '../groceries/grocery_screen.dart';
 
 class OfferBannerView extends StatelessWidget {
-  final List<String> images = [
+  const OfferBannerView({Key? key}) : super(key: key);
+
+  static const List<String> images = [
     'assets/images/banner1.jpg',
     'assets/images/banner2.jpg',
     'assets/images/banner3.jpg',
@@ -27,7 +29,7 @@ class OfferBannerView extends StatelessWidget {
           itemHeight: 100,
           duration: 500,
           itemWidth: double.infinity,
-          pagination: SwiperPagination(),
+          pagination: const SwiperPagination(),
           itemCount: images.length,
           itemBuilder: (BuildContext context, int index) => Image.asset(
             images[index],
@@ -42,7 +44,7 @@ class OfferBannerView extends StatelessWidget {
         Navigator.push(
           context,
           MaterialPageRoute(
-            builder: (context) => GroceryScreen(),
+            builder: (context) => const GroceryScreen(),
           ),
         );
       },
