@@ -8,6 +8,8 @@ import 'package:swiggy_ui/widgets/mobile/food_list_item_view.dart';
 import '../restaurants/restaurant_detail_screen.dart';
 
 class OffersScreen extends StatelessWidget {
+  const OffersScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return DefaultTabController(
@@ -16,9 +18,10 @@ class OffersScreen extends StatelessWidget {
         appBar: AppBar(
           title: Text(
             'OFFERS',
-            style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 17.0),
+            style:
+                Theme.of(context).textTheme.subtitle2!.copyWith(fontSize: 17.0),
           ),
-          bottom: TabBar(
+          bottom: const TabBar(
             indicatorColor: Colors.black,
             tabs: <Widget>[
               Tab(child: Text('RESTAURANT OFFERS')),
@@ -40,21 +43,26 @@ class OffersScreen extends StatelessWidget {
 class _RestaurantOfferView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final foods = SpotlightBestTopFood.getPopularAllRestaurants();
-    foods.addAll(SpotlightBestTopFood.getPopularAllRestaurants());
-    foods.addAll(SpotlightBestTopFood.getPopularAllRestaurants());
+    final foods = [
+      ...SpotlightBestTopFood.popularAllRestaurants,
+      ...SpotlightBestTopFood.popularAllRestaurants,
+      ...SpotlightBestTopFood.popularAllRestaurants
+    ];
 
     return Container(
       margin: const EdgeInsets.all(10.0),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
-          UIHelper.verticalSpaceSmall(),
+          UIHelper.verticalSpaceSmall,
           Text(
             'All Offers (18)',
-            style: Theme.of(context).textTheme.headline6.copyWith(fontWeight: FontWeight.bold, fontSize: 19.0),
+            style: Theme.of(context)
+                .textTheme
+                .headline6!
+                .copyWith(fontWeight: FontWeight.bold, fontSize: 19.0),
           ),
-          UIHelper.verticalSpaceMedium(),
+          UIHelper.verticalSpaceMedium,
           Expanded(
             child: ListView.builder(
               shrinkWrap: true,
@@ -66,7 +74,7 @@ class _RestaurantOfferView extends StatelessWidget {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => RestaurantDetailScreen(),
+                        builder: (context) => const RestaurantDetailScreen(),
                       ),
                     );
                   },
@@ -88,78 +96,84 @@ class _PaymentOffersCouponView extends StatelessWidget {
   Widget build(BuildContext context) {
     final coupons = AvailableCoupon.getAvailableCoupons();
 
-    return Container(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: <Widget>[
-          Container(
-            alignment: Alignment.centerLeft,
-            padding: const EdgeInsets.only(left: 15.0),
-            height: 40.0,
-            color: Colors.grey[200],
-            child: Text('Available Coupons', style: Theme.of(context).textTheme.subtitle2),
-          ),
-          Expanded(
-            child: ListView.separated(
-              shrinkWrap: true,
-              itemCount: coupons.length,
-              separatorBuilder: (context, index) => Divider(),
-              itemBuilder: (context, index) => Container(
-                margin: const EdgeInsets.all(10.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Container(
-                      padding: const EdgeInsets.all(10.0),
-                      decoration: BoxDecoration(
-                        color: Colors.orange[100],
-                        border: Border.all(color: Colors.grey[400]),
-                      ),
-                      child: Row(
-                        mainAxisSize: MainAxisSize.min,
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: <Widget>[
-                          Image.asset(
-                            'assets/images/food1.jpg',
-                            height: 10.0,
-                            width: 10.0,
-                            fit: BoxFit.cover,
-                          ),
-                          UIHelper.horizontalSpaceMedium(),
-                          Text(coupons[index].coupon, style: Theme.of(context).textTheme.subtitle2)
-                        ],
-                      ),
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: <Widget>[
+        Container(
+          alignment: Alignment.centerLeft,
+          padding: const EdgeInsets.only(left: 15.0),
+          height: 40.0,
+          color: Colors.grey[200],
+          child: Text('Available Coupons',
+              style: Theme.of(context).textTheme.subtitle2),
+        ),
+        Expanded(
+          child: ListView.separated(
+            shrinkWrap: true,
+            itemCount: coupons.length,
+            separatorBuilder: (context, index) => const Divider(),
+            itemBuilder: (context, index) => Container(
+              margin: const EdgeInsets.all(10.0),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  Container(
+                    padding: const EdgeInsets.all(10.0),
+                    decoration: BoxDecoration(
+                      color: Colors.orange[100],
+                      border: Border.all(color: Colors.grey[400]!),
                     ),
-                    UIHelper.verticalSpaceSmall(),
-                    Text(
-                      coupons[index].discount,
-                      style: Theme.of(context).textTheme.subtitle2,
+                    child: Row(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: <Widget>[
+                        Image.asset(
+                          'assets/images/food1.jpg',
+                          height: 10.0,
+                          width: 10.0,
+                          fit: BoxFit.cover,
+                        ),
+                        UIHelper.horizontalSpaceMedium,
+                        Text(coupons[index].coupon,
+                            style: Theme.of(context).textTheme.subtitle2)
+                      ],
                     ),
-                    UIHelper.verticalSpaceMedium(),
-                    CustomDividerView(
-                      dividerHeight: 1.0,
-                      color: Colors.grey,
+                  ),
+                  UIHelper.verticalSpaceSmall,
+                  Text(
+                    coupons[index].discount,
+                    style: Theme.of(context).textTheme.subtitle2,
+                  ),
+                  UIHelper.verticalSpaceMedium,
+                  const CustomDividerView(
+                    dividerHeight: 1.0,
+                    color: Colors.grey,
+                  ),
+                  UIHelper.verticalSpaceMedium,
+                  Text(
+                    coupons[index].desc,
+                    style: Theme.of(context)
+                        .textTheme
+                        .bodyText1!
+                        .copyWith(fontSize: 13.0),
+                  ),
+                  UIHelper.verticalSpaceMedium,
+                  InkWell(
+                    child: Text(
+                      '+ MORE',
+                      style: Theme.of(context)
+                          .textTheme
+                          .subtitle2!
+                          .copyWith(color: Colors.blue),
                     ),
-                    UIHelper.verticalSpaceMedium(),
-                    Text(
-                      coupons[index].desc,
-                      style: Theme.of(context).textTheme.bodyText1.copyWith(fontSize: 13.0),
-                    ),
-                    UIHelper.verticalSpaceMedium(),
-                    InkWell(
-                      child: Text(
-                        '+ MORE',
-                        style: Theme.of(context).textTheme.subtitle2.copyWith(color: Colors.blue),
-                      ),
-                    )
-                  ],
-                ),
+                  )
+                ],
               ),
             ),
-          )
-        ],
-      ),
+          ),
+        )
+      ],
     );
   }
 }
