@@ -5,10 +5,11 @@ import 'package:swiggy_ui/views/mobile/swiggy/restaurants/restaurant_detail_scre
 import 'package:swiggy_ui/widgets/responsive.dart';
 
 class PopularBrandsView extends StatelessWidget {
-  final brands = PopularBrands.getPopularBrands();
+  const PopularBrandsView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    const brands = PopularBrands.popularBrands;
     final isTabletDesktop = Responsive.isTabletDesktop(context);
 
     return Container(
@@ -16,7 +17,7 @@ class PopularBrandsView extends StatelessWidget {
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          UIHelper.verticalSpaceSmall(),
+          UIHelper.verticalSpaceSmall,
           _buildPopularHeader(context),
           LimitedBox(
             maxHeight: 170.0,
@@ -33,7 +34,8 @@ class PopularBrandsView extends StatelessWidget {
                           Navigator.push(
                             context,
                             MaterialPageRoute(
-                              builder: (context) => RestaurantDetailScreen(),
+                              builder: (context) =>
+                                  const RestaurantDetailScreen(),
                             ),
                           );
                         },
@@ -45,7 +47,7 @@ class PopularBrandsView extends StatelessWidget {
                         Container(
                           decoration: BoxDecoration(
                             border: Border.all(
-                              color: Colors.grey[300],
+                              color: Colors.grey[300]!,
                               width: 3.0,
                             ),
                             borderRadius: BorderRadius.circular(40.0),
@@ -59,15 +61,21 @@ class PopularBrandsView extends StatelessWidget {
                             ),
                           ),
                         ),
-                        UIHelper.verticalSpaceSmall(),
+                        UIHelper.verticalSpaceSmall,
                         Text(
                           brands[index].name,
-                          style: Theme.of(context).textTheme.subtitle2.copyWith(fontWeight: FontWeight.w500),
+                          style: Theme.of(context)
+                              .textTheme
+                              .subtitle2!
+                              .copyWith(fontWeight: FontWeight.w500),
                         ),
-                        UIHelper.verticalSpace(2.0),
+                        const SizedBox(height: 2.0),
                         Text(
                           brands[index].minutes,
-                          style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey, fontSize: 13.0),
+                          style: Theme.of(context)
+                              .textTheme
+                              .bodyText1!
+                              .copyWith(color: Colors.grey, fontSize: 13.0),
                         )
                       ],
                     ),
@@ -81,20 +89,22 @@ class PopularBrandsView extends StatelessWidget {
     );
   }
 
-  Container _buildPopularHeader(BuildContext context) => Container(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            Text(
-              'Popular Brands',
-              style: Theme.of(context).textTheme.headline4.copyWith(fontSize: 20.0),
-            ),
-            UIHelper.verticalSpaceExtraSmall(),
-            Text(
-              'Most ordered from around your locality',
-              style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey),
-            ),
-          ],
-        ),
+  Widget _buildPopularHeader(BuildContext context) => Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          Text(
+            'Popular Brands',
+            style:
+                Theme.of(context).textTheme.headline4!.copyWith(fontSize: 20.0),
+          ),
+          UIHelper.verticalSpaceExtraSmall,
+          Text(
+            'Most ordered from around your locality',
+            style: Theme.of(context)
+                .textTheme
+                .bodyText1!
+                .copyWith(color: Colors.grey),
+          ),
+        ],
       );
 }
