@@ -1,21 +1,20 @@
 import 'package:flutter/material.dart';
 
-import 'cart_view.dart';
 import 'home_view.dart';
 import 'menu_view.dart';
 
 class TabScreen extends StatelessWidget {
+  const TabScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Container(
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            MenuView(isTab: true, expandFlex: 1),
-            HomeView(expandFlex: 5),
-          ],
-        ),
+      body: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: const [
+          MenuView(isTab: true, expandFlex: 1),
+          HomeView(expandFlex: 5),
+        ],
       ),
     );
   }
